# swagger-pierre-le-developpeur.yaml
openapi: 3.0.3
info:
  title: Swagger Pierre le developpeur - OpenAPI 3.0
  description: |-
    Some useful links:
    - [Website](https://pierre-le-developpeur.com)
  termsOfService: http://swagger.io/terms/
  contact:
    email: contact@pierre-le-developpeur.com
  license:
    name: Apache 2.0 ?
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.2.0
...
servers:
  - url: https://pierre_le_developpeur.com/api/v0
...
paths:
...
  /user/sign_up:
    post:
      tags:
        - user
      summary: sign_up
      description: sign_up
      operationId: sign_up
      requestBody:
        description: sign_up
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Login"
        '404':
          description: Data not found
...
components:
  schemas:
...
    User:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    Login:
      type: object
      properties:
        token:
          type: string
...
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header